backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
neo4j==5.25.0
anthropic==0.57.1
cohere==5.11.0
python-dotenv==1.0.1
pydantic==2.10.2
numpy>=1.26.0
